Move reputation code to separate file
